Put offerTemplate in a separate offerTemplate file
