+++ frontend/src/assets/Accueil.module.css
@@ -69,6 +69,7 @@ p {
 }
 .favoriteIcon {
   height: 1rem;
+  background-color: red;
 }
 
 /* -------------------------oscar-------------------------*/
